Adapting to static libaray changes for hipcc
Since the introduction of BUILD_SHARED_LIBS there is only one target created : libamdhip64
Which is static ( libamdhip64.a ) or Dynamic ( libamdhip64.so ) depending on whichever it is
being compiled for. After this change we do not have any target libamdhip64_static which is
mentioned in the hipcc for linking. Hence cleaning the same.

Change-Id: Ia45c145e17b111b3f6a32299a11962b10f1cc539


[ROCm/hip commit: 9f3cf75713a14ea54e6edcc13e4ef34c33266923]