rename abort in nvcc path
[background] There is currently a compilation issue when both hip/hip_runtime.h and hip/hip_cooperative_groups.h included
in a file and hipcc used to compile it on NV platform. The issue is that an abort is defined in hip/nvcc_detail/hip_runtime.h
and it is also defined in the CUDA cooperative groups header (/cuda/include/cooperative_groups/details/helpers.h).
this is problematic and leads to a compilation issue in hipcc on NV platform.

Change-Id: I2ab6982ac4103822a1a4a0ced942cd604d6c19c1


[ROCm/clr commit: 35a3304acc0b5cba2f3f3ea227c37a8344047207]