SWDEV-273235 - Initial support for Windows CMake
This change is tied to the ROCclr cmake rework. ROCclr will now be built
as part of OpenCL itself.

All the dependencies on comgr and rocr have been moved into ROCclr
itself. This allows OpenCL to pull in all the required compiler/runtime
dependencies by simply linking against ROCclr. Note that ROCclr needs to
be configured from OpenCL cmake command line. If not, we will default to
LC + ROCr support.

Change-Id: I5271302d073aef8cfa617828d218c856711bf9ba


[ROCm/clr commit: 1c166906170ad6da158a50af61cd7dfc1a971e62]